requirements.txt
Flask==1.1.1
google-cloud-datastore==1.10.0
google-auth==1.6.3
requests==2.22.0
pymongo==3.9.0
dnspython==1.16.0
flask_googlemaps=0.2.6